Agentic_Investing_Project/requirements.txt
pandas
numpy
boto3
fastmcp